--- components/home-header.js
@@ -1,10 +1,0 @@
-export default () => <>
-  <h1>
-    <img src='https://badgen.net/static/badgen-logo.svg' /> Badgen
-  </h1>
-  <p>Fast badge generating service.</p>
-  <p>
-    <a href='https://badgen.net'>classic</a>
-    <a href='https://flat.badgen.net'>flat</a>
-  </p>
-</>
